+++ transpiler/base.py
@@ -1,6 +1,5 @@
-from functools import lru_cache
 import re
-from enum import Enum, EnumMeta
+from enum import Enum
 
 
 WHITESPACE = re.compile('\S')
@@ -10,93 +9,23 @@ class TranspilerError(Exception):
     pass
 
 
-class EntityMeta(EnumMeta):
-    @lru_cache
-    def __contains__(cls, item):
-        try:
-            cls(item)
-        except ValueError:
-            return False
-        else:
-            return True
-
-
-
-class Entity(Enum, metaclass=EntityMeta):
+class Entity(Enum):
     def __repr__(self) -> str:
         return self.value
 
 
-class Tag(Entity):
-    """
-    Possible token names enumeration.
-    """
+class Special(Entity):
+    LAMBDA = '__LAMBDA__'
+    START = '__START__'
+    LIMITER = '__LIMITER__'
 
-    # general
-    ID = 'ID'
-    NUMBER_INT = 'NUMBER_INT'
-    NUMBER_FLOAT = 'NUMBER_FLOAT'
-    LBRACKET = 'LBRACKET'
-    RBRACKET = 'RBRACKET'
-    LBRACKET_SQUARE = 'LBRACKET_SQUARE'
-    RBRACKET_SQUARE = 'RBRACKET_SQUARE'
-    SEMICOLON = 'SEMICOLON'
-    COLON = 'COLON'
-    COMMA = 'COMMA'
-    DOT = 'DOT'
-    QUOTE = 'QUOTE'
-    DQUOTE = 'DQOUTE'
-    LAMBDA = 'LAMBDA'
 
-    # types
-    T_INTEGER = 'T_INTEGER'
-    T_REAL = 'T_REAL'
-    T_BOOLEAN = 'T_BOOLEAN'
-    T_CHAR = 'T_CHAR'
-    T_STRING = 'T_STRING'
-    T_ARRAY = 'T_ARRAY'
+class Terminal(Entity):
+    pass
 
-    # comparisons
-    EQ = 'EQ'
-    NE = 'NE'
-    LE = 'LE'
-    LT = 'LT'
-    GE = 'GE'
-    GT = 'GT'
 
-    # operators
-    PLUS = 'PLUS'
-    MINUS = 'MINUS'
-    MULTIPLY = 'MULTIPLY'
-    DIVIDE = 'DIVIDE'
-    ASSIGN = 'ASSIGN'
-    PLUS_ASSIGN = 'PLUS_ASSIGN'
-    MINUS_ASSIGN = 'MINUS_ASSIGN'
-    MULTIPLY_ASSIGN = 'MULTIPLY_ASSIGN'
-    DIVIDE_ASSIGN = 'DIVIDE_ASSIGN'
-    RANGE = 'RANGE'
-
-    # boolean
-    TRUE = 'TRUE'
-    FALSE = 'FALSE'
-
-    # other key words
-    VAR = 'VAR'
-    IF = 'IF'
-    THEN = 'THEN'
-    ELSE = 'ELSE'
-    CASE = 'CASE'
-    OF = 'OF'
-    FOR = 'FOR'
-    WHILE = 'WHILE'
-    REPEAT = 'REPEAT'
-    UNTIL = 'UNTIL'
-    DO = 'DO'
-    TO = 'TO'
-    BEGIN = 'BEGIN'
-    END = 'END'
-    PROCEDURE = 'PROCEDURE'
-    FUNCTION = 'FUNCTION'
+class NonTerminal(Entity):
+    pass
 
 
 class LexerRule:
@@ -104,7 +33,7 @@ class LexerRule:
     Token definition and regular expression mapping. Used by lexer.
     """
 
-    def __init__(self, tag: Tag, regex: str) -> None:
+    def __init__(self, tag: Terminal, regex: str) -> None:
         self.tag = tag
         self.regex = regex
 
@@ -114,7 +43,7 @@ class Token:
     Minimal sensible unit of code sequence.
     """
 
-    def __init__(self, tag: Tag, value: str, pos: int):
+    def __init__(self, tag: Terminal, value: str, pos: int):
         self.tag = tag
         self.value = value
         self.pos = pos
@@ -122,22 +51,10 @@ class Token:
     def __str__(self):
         return f'{self.value} {self.tag.value}'
 
-
-class NonTerm(Entity):
-    _START = '_START'
-    DESCR = 'DESCR'
-    PROG = 'PROG'
-    VARS = 'VARS'
-    EXPR = 'EXPR'
-    TYPE = 'TYPE'
-    CALL = 'CALL'
-    ARGS = 'ARGS'
-
-
 class GrammarRule:
     def __init__(
         self,
-        left: NonTerm,
+        left: NonTerminal,
         right: set[tuple()]
     ):
         self.left = left
